Just Start Somewhere
I'm getting a little burned out on the book, so I'm pivoting to a book I've
wanted to do for a while called Mazes for Programmers. The book is written with
Ruby code, but I'm translating that to C++. I've already hit several 'road
blocks'.

For example, I don't think I'm using "this" correctly. I'm not sure how to
delete elements from unordered maps. I'm not sure how unordered (or ordered)
maps work. I'm not sure how to pass classes around to other classes. I probably
need to use references and pointers here? It's going to be slow going.

The book starts off with ASCII mazes, which is a relief. This code does not
work, however I wanted to start something of my own before leaving for the day.