Grid Life
During the Hour of Hard Things today I worked on making this grid. I had tried
previously to get a grid and a maze going in C++, but was immediately stumped by
how c-stye strings work and also the trickiness of two dimensional arrays when
working with pointers.

Last night as I was reading my C++ book he walked through making a
'SimpleString' class and that really unlocked some doors. Once I realized that I
was just going to have to make my own SimpleString class with it's own 'append'
and 'append_line' functions everything else started to make more
sense. Additionally I let go of the two demensional array in favor of a single
array and using two (actually 3) for loops to iterate over it (and then adding
the col and row together to get the index of the cell I was working with).

It was also very helpful to have worked out the problem in C# previously so I
wasn't thinking about the langauge and the problem. I could just focus on the
new things of the language.

I still need to make a few more things dynamic (such as the size of the maze and
the number of rows and columns and hwo the loops work). But I'm really pleased
with how it looks and I think because I added in a check for the specific cell,
making connected cells will fit in nicely.