Rebuilt Binary Tree Maze!
So after cleaning the slate I rebuilt the game and it only took 3
weeks!

I'm really pleased with the design this time. I feel like it maybe
takes advantage of pointers more? Each pair of neighboring cells point
to the same "connection" object. So for example one cells North member
points to a connection object that is also pointed to by the above
cells "South" connection type.

This is done here:
	      Connection* conn = new Connection{ConnectionType::Wall};
+	      cells[r][c].North = conn;
+	      cells[r-1][c].South = conn;

and here:

	  if(c != cols-1)
+	    {
+	      Connection* conn = new Connection{ConnectionType::Wall};
+	      cells[r][c].East = conn;
+	      cells[r][c+1].West = conn;
+	    }

This means that when i change the south wall for one cell I don't have
to worry about finding the cell below and changin the "north"
wall. Because it's already changed because they point to the same
object.

This behavior isn't novel to C++, but I think because pointers are so
explicit it has me thinking along these lines.

Now that I have a pretty decent grid that I can easily manipulate I'm
excited to see what other types of mazes I can get up to making.